reviewed project layout for React Query Refactor
